feat/ref: file path error handling + clean code

+++ scripts/leon.py
@@ -6,11 +6,9 @@ from skimage.metrics import structural_similarity as ssim
 import numpy as np
 import pandas as pd
 import imagehash
-import keras
 import cv2
 from keras import layers
 from keras.models import Sequential  # type: ignore
-
 import matplotlib.pyplot as plt
 from scripts.styler import Styler
 
@@ -43,6 +41,9 @@ class Leon:
         Returns:
             list of PIL.Image.Image: List of Image objects containing the images from the ZIP file.
         """
+        if not os.path.exists(path):
+            raise FileNotFoundError(f"File not found: {path}")
+
         # Get the directory where the ZIP file will be extracted
         extract_dir = os.path.splitext(path)[0]
 
@@ -77,6 +78,9 @@ class Leon:
         Returns:
             list of PIL.Image.Image: List of Image objects containing the images.
         """
+        if not os.path.exists(path):
+            raise FileNotFoundError(f"Directory not found: {path}")
+
         images = []
         count = 0
 
@@ -122,6 +126,8 @@ class Leon:
             list of str: List of perceptual hash strings.
             list of PIL.Image.Image: List of duplicate images.
         """
+        if not os.path.exists(path):
+            raise FileNotFoundError(f"Directory not found: {path}")
 
         # Dictionary to store the hashes of the images
         hashes = {}
@@ -208,7 +214,7 @@ class Leon:
         image_path,
         output_dir,
         df_train,
-        num_images=3,
+        num_images=2,
         rotation=0.5,
         contrast=0.5,
     ):
@@ -224,28 +230,39 @@ class Leon:
             rotation (float): The maximum rotation angle in degrees. Default is 0.5.
             contrast (float): The maximum contrast factor. Default is 0.5.
         """
-        # Initialize df_train as an empty DataFrame if not provided
+        # Error handling
+        if not os.path.exists(image_path):
+            raise FileNotFoundError(f"File not found: {image_path}")
+        if not os.path.exists(output_dir):
+            os.makedirs(output_dir)
         if df_train is None:
             raise ValueError("df_train is required.")
 
+        # Data augmentation transformations
         data_augmentation = Sequential(
             [
-                layers.RandomFlip(),
-                layers.RandomRotation(rotation),
-                layers.RandomContrast(contrast),
+                layers.RandomFlip(),  # Randomly flip the image horizontally or vertically
+                layers.RandomRotation(rotation),  # Randomly rotate the image
+                layers.RandomContrast(
+                    contrast
+                ),  # Randomly adjust the contrast of the image
             ]
         )
 
+        # Read the image
         img = cv2.imread(image_path)
 
+        # Convert the image to a numpy array and add a batch dimension
         img_array = np.expand_dims(img, axis=0)
 
+        # Update the dataframe
         styler.boxify(f"Augmenting image: {image_path}")
 
         image_filename = os.path.basename(image_path)
 
         path_parts = image_path.split(os.path.sep)
 
+        # Extract the category and style from the image path
         if len(path_parts) >= 5:
             category = path_parts[-2]
             style = path_parts[-3]
@@ -300,17 +317,24 @@ class Leon:
 
                 pd.DataFrame: A DataFrame containing the image paths, classes, styles, widths, and heights.
         """
+        if not os.path.exists(dir):
+            raise FileNotFoundError(f"Directory not found: {dir}")
 
+        # Initialize the dictionary to store the data
         data_dict = {
             "Path": [],
             "Category": [],
             "Style": [],
             "Width": [],
             "Height": [],
+            "MinValue": [],
+            "MaxValue": [],
+            "StdDev": [],
         }
 
         data_dir = os.path.relpath(dir)
 
+        # Get the categories and styles
         categories = [
             folder
             for folder in os.listdir(data_dir)
@@ -318,6 +342,7 @@ class Leon:
         ]
 
         for category in categories:
+            # Get the styles for each category
             category_dir = os.path.join(data_dir, category)
             styles = [
                 folder
@@ -325,8 +350,11 @@ class Leon:
                 if os.path.isdir(os.path.join(category_dir, folder))
             ]
 
+            # Iterate over the styles
             for style in styles:
                 style_dir = os.path.join(category_dir, style)
+
+                # Iterate over the images in the style directory
                 for file in os.listdir(style_dir):
                     img_path = os.path.join(style_dir, file)
                     data_dict["Path"].append(img_path)
@@ -339,6 +367,19 @@ class Leon:
                             width, height = img.size
                             data_dict["Width"].append(width)
                             data_dict["Height"].append(height)
+
+                            # Convert image to numpy array
+                            img_array = np.array(img)
+
+                            # Compute statistics
+                            min_val = np.min(img_array)
+                            max_val = np.max(img_array)
+                            std_dev = np.std(img_array)
+
+                            # Add statistics to the dictionary
+                            data_dict["MinValue"].append(min_val)
+                            data_dict["MaxValue"].append(max_val)
+                            data_dict["StdDev"].append(std_dev)
                     except Exception as e:
                         print(f"Error processing image '{img_path}': {e}")
 
@@ -353,6 +394,9 @@ class Leon:
             width (int): The width of the resized image.
             height (int): The height of the resized image.
         """
+        if not os.path.exists(path):
+            raise FileNotFoundError(f"File not found: {path}")
+        
         # Open the image file
         with open(path, "rb") as f:
             image = Image.open(f)
@@ -363,24 +407,41 @@ class Leon:
             # Save the resized image
             resized_image.save(path)
 
-    def normalise_pixel_values(self, image_path):
+    def normalize_image(self, image_path, min_value=0, max_value=1):
         """
-        Normalises the pixel values of the image to the range [0, 1].
+        Normalize the pixel values of an image to the range [0, 1].
 
         Parameters:
-            image (PIL.Image.Image): The image to normalise.
+            image_path (str): The path to the image file.
+            min_value (float): The minimum value of the normalized range. Default is 0.
+            max_value (float): The maximum value of the normalized range. Default is 1.
 
         Returns:
-            np.ndarray: The normalised image as a NumPy array.
+            str: The path to the normalized image.
         """
-        # Convert the image to a NumPy array
-        image = Image.open(image_path)
-        image_array = np.array(image)
+        if not os.path.exists(image_path):
+            raise FileNotFoundError(f"File not found: {image_path}")
 
-        # Normalise the pixel values to the range [0, 1]
-        normalised_image = image_array / 255.0
+        if not isinstance(min_value, (int, float)) or not isinstance(
+            max_value, (int, float)
+        ):
+            raise TypeError("min_value must be an integer or float")
 
-        return normalised_image
+        if min_value >= max_value:
+            raise ValueError("min_value must be less than max_value")
+
+        styler.boxify(f"Normalizing image: {image_path}")
+        # Open the image
+        image = cv2.imread(image_path)
+
+        # Normalize the pixel values to the range [0, 1]
+        normalized_image = cv2.normalize(
+            image, None, min_value, max_value, cv2.NORM_MINMAX, dtype=cv2.CV_32F
+        )
+
+        print(">>> Processed successfully")
+
+        return normalized_image
 
     def remove_folder(self, path):
         """
@@ -389,4 +450,98 @@ class Leon:
         Parameters:
             path (str): The path to the folder to remove.
         """
+        if not os.path.exists(path):
+            raise FileNotFoundError(f"Folder not found: {path}")
+
+        print(
+            "This is a destructive operation as folder will be deleted permanently. Are you sure you want to continue? (y/n)"
+        )
+
+        response = input()
+
+        if response.lower() != "y" and response.lower() != "yes":
+            print("Operation cancelled.")
+            return
+
+        print(f"Removing folder: {path}")
         shutil.rmtree(path)
+
+    # Remove output folders and files generated by the previous run
+
+    def remove_nonraw_files(self, img_paths=[]):
+        """
+        Removes output folders and files generated by the previous run.
+
+        Parameters:
+            img_paths (list[str]): List of image paths in the training set.
+        """
+        # Check if img_paths is provided
+        if not img_paths:
+            raise ValueError("img_paths is required.")
+
+        # Confirm with the user before proceeding
+        print(
+            "This is a destructive operation as files will be deleted permanently. Are you sure you want to continue? (y/n)"
+        )
+
+        response = input()
+
+        if response.lower() != "y" and response.lower() != "yes":
+            print("Operation cancelled.")
+            return
+
+        # Initialize set to store processed directories
+        processed_directories = set()
+
+        # Get all image path in the training set
+        for img_path in img_paths:
+            # Extract directory name
+            directory = os.path.dirname(img_path)
+
+            # Check if directory has already been processed
+            if directory in processed_directories:
+                continue
+
+            # Add directory to the set of processed directories
+            processed_directories.add(directory)
+
+            # Remove files starting with "augmented_" if they exist
+            files_in_directory = os.listdir(directory)
+            for file_name in files_in_directory:
+                # Check if file name starts with "aug_"
+                if file_name.startswith("aug_") or file_name.endswith("_norm.jpg"):
+                    # Construct file path
+                    file_path = os.path.join(directory, file_name)
+
+                    # Remove file
+                    try:
+                        os.remove(file_path)
+                        print(f"Removed file: {file_path}")
+                    except OSError as e:
+                        print(f"Error removing file: {file_path}, {e}")
+
+    def get_image_paths(self, directory):
+        """
+        Get the paths of all image files in the directory and its subdirectories.
+
+        Parameters:
+            directory (str): The directory to search for image files.
+
+        Returns:
+            list of str: List of image paths.
+        """
+        if not os.path.exists(directory):
+            raise FileNotFoundError(f"Directory not found: {directory}")
+
+        image_paths = []
+        # Walk through the directory and its subdirectories
+        for root, dirs, files in os.walk(directory):
+            # Iterate over the files in the current directory
+            for file in files:
+                # Check if the file is an image (you can modify this condition as needed)
+                if file.endswith((".jpg")):
+                    # Construct the full path to the image file
+                    image_path = os.path.join(root, file)
+                    # Add the image path to the list
+                    image_paths.append(image_path)
+        return image_paths
